requirements.txt
# Requirements optimized for Streamlit deployment with accordion support
# Core LLM & RAG Frameworks
langchain>=0.2.12
langchain-community>=0.2.10
langchain-core==1.0.2
faiss-cpu>=1.8.0
openai>=1.12.0
langchain-openai
tiktoken

# Streamlit UI with deployment optimizations
streamlit>=1.36.0
streamlit-accordion>=0.1.0

# Data Handling & Formats
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
openpyxl>=3.1.2
xlrd>=2.0.1
fastparquet>=2024.2.0

# File & Config
PyYAML>=6.0.2
tqdm>=4.66.0

# API Integration
requests>=2.32.0

# Google Generative AI (Compatible versions)
langchain-google-genai==3.0.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# For ZIP handling and generic IO
python-magic>=0.4.27

# Lightweight visualization (better for deployment)
matplotlib>=3.9.0

# HuggingFace (for offline embedding models)
transformers>=4.41.0
huggingface-hub>=0.18.1
langchain-huggingface>=0.0.1
sentence-transformers>=2.7.0